# perf(serialize): faster serialization of typed arrays and objects with string keys