--- a/src/serialize.ts
+++ b/src/serialize.ts
@@ -22,6 +22,10 @@ const Serializer = /*@__PURE__*/ (function () {
     #context = new Map();
 
     compare(a: any, b: any): number {
+      if (typeof a === "string" && typeof b === "string") {
+        return a.localeCompare(b);
+      }
+
       if (typeof a === "number" && typeof b === "number") {
         return a - b;
       }