@zsilbi 2025-03-03
`compare()` could be improved further for mixed key objects/sets if we would prioritize numbers, but this would break the current order.

For example:

```scala
Expected: ""{a:Set[#1,1],b:Set[#1,1]}""
Received: ""{a:Set[1,#1],b:Set[1,#1]}""
```

Possible optimization:
```ts
compare(a: any, b: any): number {
  if (typeof a === "string" && typeof b === "string") {
    return a.localeCompare(b);
  }

  if (typeof a === "number") {
    if (typeof b === "number") {
      return a - b;
    }
    return -1;
  }
  if (typeof b === "number") {
    return 1;
  }

  return String.prototype.localeCompare.call(
    this.serialize(a, true),
    this.serialize(b, true),
  );
}
```