@zsilbi 2025-03-03
Overall benchmark with all presets: 

```scala
 ✓ test/benchmarks.bench.ts > benchmarks > all presets 1256ms
     name                  hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · serialize @ main  175.20  4.9805  9.1068  5.7078  5.7614  9.1068  9.1068  9.1068  ±2.69%       88
   · serialize         185.85  4.8704  6.3206  5.3807  5.5521  6.3206  6.3206  6.3206  ±1.22%       93   fastest

 BENCH  Summary

  serialize - test/benchmarks.bench.ts > benchmarks > all presets
    1.06x faster than serialize @ main
 ```